[js] Avoid optimizing Std.is away in api_inline. (#12133)
This allows deprecation checks to run.
Codegen does not change as Std.is is inline and calls Std.isOfType
which is still special-cased.
(Std.isOfType itself still needs to be special-cased despite calling js.Boot.__instanceof because otherwise exception codegen changes slightly)
Added a codegen test and a test to verify the deprecation warning triggers.