[cpp] Do not generate toString in __scriptableFunctions (#12375)
* [cpp] Fix toString breaking scriptable mapping

The toString method is filtered out in the vtable mapping on line 973
and is already present in the `__scriptableFunctions` of hx::Object.
We need to filter it out here too to ensure a correct vtable can be
constructed from merging all `__scriptableFunctions` arrays.

* [tests] Add case for toString scriptable issue

* [tests] Clean up

Remove unnecessary public modifier

Fix order of expected vs actual equality values

* [cpp] Add comments to clarify toString handling

* [tests] Add issue number as suffix to avoid clash