[hxb] Recover from BadModule failure (#12395)
This currently can happen when current display module (which will be
invalidated in case of completion for example) is an indirect dependency
of a module that ends up being loaded from init macros. Other cases
should usually behave already from what I've seen.

So here, when we're _not_ in full restore mode (display requests only,
with some additional constraints), we're allowing hxb restoration of
"BadModule" modules, avoiding a compiler failure. That "bad" module is
then discarded, avoiding cache corruption (yes, this means it
technically can happen more than once in a request, but that should be
_very_ rare and not that much of an issue since no typing is involved).