[tests] Fix misc issues in cpp tests (#12472)
* [tests] Enable cpp thread tests on linux arm64

* [tests] Use default architecture for cpp tests

Windows shouldn't be built with x86_32 anymore, since most users will be
using x86_64.

In CI we always want the host architecture anyway. For local runs, users
can just pass `-D HXCPP_ARCH=...` as cli args when running RunCi.

* [tests] Add cli args to all cpp builds

* Revert "[tests] Add cli args to all cpp builds"

This reverts commit 905cf81d82e314ee6dcc1955b0cd19f66ba05297.

This commit fixed argument passing to sys tests, which caused a test to
run that wasn't running before which is currently broken on windows.

Reverting for now, this will be handled in #12473