[cpp] Mark overridden scriptable methods (#12413)
* [cpp] Include marked override scriptable methods

Rather than omitting method overrides, we now include them but mark them
as overrides to prevent them from being added as separate slots in the
vtable. This keeps vtable indices working correctly while keeping
super.method calls working as they should.

* [tests] Add test for cppia super.method calls

* [cpp] Remove HXCPP_CPPIA_SUPER_ARG macro usage

It is not needed as it is always defined the same way by hxcpp. We can
instead just pass the argument directly, or omit it if it is empty.

In some cases, HXCPP_CPPIA_SUPER_ARG(0) was being passed to the
inIsStatic argument.

* [cpp] Remove 0 arguments to ScriptNamedFunction

They are optional in the api version of hxcpp we require, and have been
optional for a very long time.